@@ -24,6 +24,7 @@
     "gulp-concat-css": "^2.3.0",
     "gulp-copy": "^1.0.0",
     "gulp-util": "3.0.7",
+    "prop-types": "^15.5.10",
     "react": "15.4.2",
     "react-dom": "15.4.2",
     "webpack": "1.13.3"
@@ -35,6 +36,6 @@
     "superagent": "3.4.1"
   },
   "peerDependencies": {
-    "react": "15.4.2"
+    "react": "^15.5.4"
   }
 }
